--- constant.py
@@ -1,12 +1,0 @@
-DIGITS = "0123456789"
-
-# token type => TT
-TT_INT = "INT"
-TT_FLOAT = "FLOAT"
-TT_PLUS = "PLUS"
-TT_MINUS = "MINUS"
-TT_MUL = "MUL"
-TT_DIV = "DIV"
-TT_LPAREN = "LPAREN"
-TT_RPAREN = "RPAREN"
-TT_EOF = "EOF"  # 终止符
